requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
lxml>=4.9.0

# PDF processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# LLM
google-generativeai>=0.3.0

# API Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Utilities
cachetools>=5.3.0
pyyaml>=6.0.0  # For config.yaml support
python-dotenv>=1.0.0  # For .env file support

# Redis support (required for caching)
redis>=5.0.0

# Optional: PostgreSQL support (install with: pip install psycopg2-binary)
# psycopg2-binary>=2.9.0
